$8 - Move GLFW interop types to another folder and namespace
