+++ Testing/VelaptorTests/OpenGL/GLFWMonitorsTests.cs
@@ -12,6 +12,7 @@ namespace VelaptorTests.OpenGL
     using Velaptor;
     using Velaptor.Hardware;
     using Velaptor.NativeInterop;
+    using Velaptor.NativeInterop.GLFW;
     using Velaptor.OpenGL;
     using Xunit;
