+++ Velaptor/OpenGL/GLFWMonitors.cs
@@ -9,6 +9,7 @@ namespace Velaptor.OpenGL
     using System.Numerics;
     using Velaptor.Hardware;
     using Velaptor.NativeInterop;
+    using Velaptor.NativeInterop.GLFW;
 
     /// <summary>
     /// Gets all of the monitors in the system.
